requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.0
paramiko==3.3.1
python-dotenv==1.0.0
docker==6.1.3